+++ VERA/VERAVonage/VERAVonageTests/Mock/MockPublisherRepository.swift
@@ -5,12 +5,16 @@
 import Foundation
 import OpenTok
 import VERADomain
+import VERATestHelpers
 import VERAVonage
 
 public class MockPublisherRepository: PublisherRepository {
 
     public func getPublisher() -> any VERAPublisher {
-        VonagePublisher(publisher: OTPublisher(delegate: nil)!)
+        VonagePublisher(
+            publisher: OTPublisher(delegate: nil)!,
+            transformerFactory: MockTransformerFactory()
+        )
     }
 
     public func resetPublisher() {
